requirements.txt
numpy==1.26.0
flask==2.2.2
Pillow==10.0.1
tensorflow-cpu==2.14.0